#10 two entrypoints added as replacement for config
